+++ pkg/action/cl001/ac001/zz_generated.executor.go
@@ -7,26 +7,20 @@ import (
 	"github.com/giantswarm/micrologger"
 	"github.com/spf13/cobra"
 
-	"github.com/giantswarm/awscnfm/v12/pkg/action"
 	"github.com/giantswarm/awscnfm/v12/pkg/config"
 )
 
 type ExecutorConfig struct {
-	Clients *action.Clients
 	Command *cobra.Command
 	Logger  micrologger.Logger
 }
 
 type Executor struct {
-	clients *action.Clients
 	command *cobra.Command
 	logger  micrologger.Logger
 }
 
 func NewExecutor(config ExecutorConfig) (*Executor, error) {
-	if config.Clients == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Clients must not be empty", config)
-	}
 	if config.Command == nil {
 		return nil, microerror.Maskf(invalidConfigError, "%T.Command must not be empty", config)
 	}
@@ -35,7 +29,6 @@ func NewExecutor(config ExecutorConfig) (*Executor, error) {
 	}
 
 	e := &Executor{
-		clients: config.Clients,
 		command: config.Command,
 		logger:  config.Logger,
 	}