+++ pkg/action/cl001/ac005/executor.go
@@ -3,22 +3,44 @@ package ac005
 import (
 	"context"
 
+	"github.com/giantswarm/k8sclient/v3/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
+
+	"github.com/giantswarm/awscnfm/v12/pkg/client"
+	"github.com/giantswarm/awscnfm/v12/pkg/config"
+	"github.com/giantswarm/awscnfm/v12/pkg/env"
 )
 
 func (e *Executor) execute(ctx context.Context) error {
-	crs, err := newCRs(ctx, e.tenantCluster, e.clients.ControlPlane.RESTConfig().Host)
+	scope := "cl001"
+	id := config.Cluster(scope, env.TenantCluster())
+
+	var err error
+
+	var cpClients k8sclient.Interface
+	{
+		c := client.ControlPlaneConfig{
+			Logger: e.logger,
+		}
+
+		cpClients, err = client.NewControlPlane(c)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	}
+
+	crs, err := newCRs(ctx, id, cpClients.RESTConfig().Host)
 	if err != nil {
 		return microerror.Mask(err)
 	}
 
 	{
-		err = e.clients.ControlPlane.CtrlClient().Create(ctx, crs.MachineDeployment)
+		err = cpClients.CtrlClient().Create(ctx, crs.MachineDeployment)
 		if err != nil {
 			return microerror.Mask(err)
 		}
 
-		err = e.clients.ControlPlane.CtrlClient().Create(ctx, crs.AWSMachineDeployment)
+		err = cpClients.CtrlClient().Create(ctx, crs.AWSMachineDeployment)
 		if err != nil {
 			return microerror.Mask(err)
 		}