+++ pkg/project/project.go
@@ -5,7 +5,12 @@ var (
 	gitSHA      = "n/a"
 	name        = "awscnfm"
 	source      = "https://github.com/giantswarm/awscnfm"
-	version     = "12.1.1-dev"
+	// version is synchronized with the latest Giant Swarm release on the major
+	// and minor level so that we know which version we use as base for
+	// conformance testing. Please make sure that at least major and minor
+	// levels of this version aligns with each new release so that we stay
+	// synchronized with the latest AWS release we publish.
+	version = "12.1.2-dev"
 )
 
 func Description() string {