+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Modified the documentation and main README.
 - Changed the default description for cluster scopes slightly, for consistent use of the "cluster scope" term.
 
+## [12.1.1] - 2020-08-12
+
 ## [12.1.0] - 2020-08-12
 
 ### Added
@@ -28,5 +30,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Include namespace to delete cluster CR's
 
-[Unreleased]: https://github.com/giantswarm/awscnfm/compare/v12.1.0...HEAD
+[Unreleased]: https://github.com/giantswarm/awscnfm/compare/v12.1.1...HEAD
+[12.1.1]: https://github.com/giantswarm/awscnfm/compare/v12.1.0...v12.1.1
 [12.1.0]: https://github.com/giantswarm/awscnfm/releases/tag/v12.1.0